total corrections in chats and other

+++ app/src/main/java/com/example/lipe/BatteryReceiver.kt
@@ -22,7 +22,7 @@ class BatteryReceiver : BroadcastReceiver() {
         if(intent?.action == Intent.ACTION_BATTERY_CHANGED) {
             val level = intent.getIntExtra(BatteryManager.EXTRA_LEVEL, -1)
             val scale = intent.getIntExtra(BatteryManager.EXTRA_SCALE, -1)
-            val battery = level * 100 / scale.toFloat()
+            val battery = level * 100 / scale
             batteryLevelLiveData.postValue(battery.toString())
         }
     }
